refactored the code to address the review comments
